OpenDCS users and roles dao extracted.
