+++ java/opendcs-implementation/src/main/java/org/opendcs/database/impl/opendcs/OpenDcsPgProvider.java
@@ -43,6 +43,7 @@ import org.opendcs.database.dai.RolesDao;
 import org.opendcs.database.dai.UserManagementDao;
 import org.opendcs.database.dai.UsersDao;
 import org.opendcs.database.impl.opendcs.dao.OpenDcsPgUserManagementImpl;
+import org.opendcs.database.impl.opendcs.dao.auth.RolesDaoImpl;
 import org.opendcs.database.impl.opendcs.jdbi.column.databasekey.DatabaseKeyArgumentFactory;
 import org.opendcs.database.impl.opendcs.jdbi.column.databasekey.DatabaseKeyColumnMapper;
 import org.opendcs.database.model.Role;
@@ -121,12 +122,13 @@ public class OpenDcsPgProvider implements MigrationProvider
         jdbi.useTransaction(h ->
         {
             var tx = new JdbiTransaction(h, context);
-            var dao = new OpenDcsPgUserManagementImpl();
+            var idpDao = new OpenDcsPgUserManagementImpl();
+            var rolesDao = new RolesDaoImpl();
 
             try(Call createUser = h.createCall("call create_user(:user,:pw)");
                 Call assignRole = h.createCall("call assign_role(:user,:role)");)
             {
-                setupIdentityProvider(tx, dao, dao);
+                setupIdentityProvider(tx, rolesDao, idpDao);
                 createUser.bind("user",username)
                           .bind("pw", password)
                           .invoke();
